BigQuery/requirements.txt
google-cloud-bigquery==3.*
google-cloud-storage==2.*

